Bomb Effect complete / EmberFire in  progress
Started EmberFire as the bomb and explosion on impact but found better imagery for EmberFire. Will need to get some scripts renamed, etc. so there's no confusion. Created a dummy script to emulate the tower/player and worked out some temp code for integrating the projectiles and the towers. Implemented in-editor control of how long it takes for a projectile to auto destroy for testing. Great progress today.