Charmander Ember working
Currently only have the ember going in one direction for testing purposes but need to add direction vectors for correct orientation as the ember moves through space is fired in different directions. Need to instantiate with correct rotation as well.
Added some more temp code in the player scripts for possible approach to combining towers and projectiles.